+++ android/expoview/src/main/java/host/exp/exponent/fcm/ExpoFcmMessagingService.java
@@ -1,12 +1,19 @@
 package host.exp.exponent.fcm;
 
-import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
-import host.exp.exponent.Constants;
-import host.exp.exponent.notifications.PushNotificationHelper;
+import java.util.Map;
 
-public class ExpoFcmMessagingService extends FirebaseMessagingService {
+import expo.modules.notifications.FirebaseListenerService;
+import expo.modules.notifications.notifications.model.NotificationContent;
+import expo.modules.notifications.notifications.model.NotificationRequest;
+import expo.modules.notifications.notifications.model.triggers.FirebaseNotificationTrigger;
+import host.exp.exponent.Constants;
+import host.exp.exponent.kernel.ExperienceId;
+import host.exp.exponent.notifications.PushNotificationHelper;
+import versioned.host.exp.exponent.modules.universal.notifications.ScopedNotificationRequest;
+
+public class ExpoFcmMessagingService extends FirebaseListenerService {
 
   @Override
   public void onNewToken(String token) {
@@ -14,6 +21,7 @@ public class ExpoFcmMessagingService extends FirebaseMessagingService {
       return;
     }
 
+    super.onNewToken(token);
     FcmRegistrationIntentService.registerForeground(getApplicationContext(), token);
   }
 
@@ -23,6 +31,19 @@ public class ExpoFcmMessagingService extends FirebaseMessagingService {
       return;
     }
 
+    super.onMessageReceived(remoteMessage);
     PushNotificationHelper.getInstance().onMessageReceived(this, remoteMessage.getData().get("experienceId"), remoteMessage.getData().get("channelId"), remoteMessage.getData().get("message"), remoteMessage.getData().get("body"), remoteMessage.getData().get("title"), remoteMessage.getData().get("categoryId"));
+  }
+
+  @Override
+  protected NotificationRequest createNotificationRequest(String identifier, NotificationContent content, FirebaseNotificationTrigger notificationTrigger) {
+    ExperienceId experienceId;
+    Map<String, String> data = notificationTrigger.getRemoteMessage().getData();
+    if (!data.containsKey("experienceId")) {
+      experienceId = null;
+    } else {
+      experienceId = ExperienceId.create(data.get("experienceId"));
+    }
+    return new ScopedNotificationRequest(identifier, content, notificationTrigger, experienceId);
   }
 }