+++ packages/expo-auth-session/src/AuthRequest.ts
@@ -163,9 +163,8 @@ export class AuthRequest {
 
     let result: WebBrowser.WebBrowserAuthSessionResult;
     try {
-      result = await WebBrowser.openAuthSessionAsync(startUrl, returnUrl, {
-        showInRecents: options.showInRecents,
-      });
+      const { useProxy, ...openOptions } = options;
+      result = await WebBrowser.openAuthSessionAsync(startUrl, returnUrl, openOptions);
     } finally {
       _authLock = false;
     }
@@ -230,7 +229,7 @@ export class AuthRequest {
       }
     }
 
-    if (request.codeChallengeMethod) {
+    if (request.usePKCE && request.codeChallengeMethod) {
       params.code_challenge_method = request.codeChallengeMethod;
     }
 
@@ -247,7 +246,10 @@ export class AuthRequest {
     params.client_id = request.clientId;
     params.response_type = request.responseType!;
     params.state = request.state;
-    params.scope = request.scopes.join(' ');
+
+    if (request.scopes.length) {
+      params.scope = request.scopes.join(' ');
+    }
 
     const query = QueryParams.buildQueryString(params);
     // Store the URL for later