+++ packages/expo-calendar/src/Calendar.ts
@@ -701,7 +701,11 @@ function stringifyIfDate(date: any): any {
 
 function stringifyDateValues(obj: object): object {
   return Object.keys(obj).reduce((acc, key) => {
-    acc[key] = stringifyIfDate(obj[key]);
+    const value = obj[key];
+    if (typeof value === 'object' && !(value instanceof Date)) {
+      return { ...acc, [key]: stringifyDateValues(value) };
+    }
+    acc[key] = stringifyIfDate(value);
     return acc;
   }, {});
 }