+++ packages/expo-font/src/Font.ts
@@ -209,4 +209,3 @@ async function unloadFontInNamespaceAsync(
 }
 
 export { FontDisplay, FontSource, FontResource, UnloadFontOptions };
-export { useFonts } from './FontHooks';