+++ packages/expo-notifications/src/NotificationScheduler.ts
@@ -1,10 +1,10 @@
 import { NativeModulesProxy, ProxyNativeModule } from '@unimodules/core';
 
 import { NotificationTriggerInput } from './NotificationScheduler.types';
-import { Notification, NotificationContentInput } from './Notifications.types';
+import { NotificationRequest, NotificationContentInput } from './Notifications.types';
 
 export interface NotificationSchedulerModule extends ProxyNativeModule {
-  getAllScheduledNotificationsAsync: () => Promise<Notification[]>;
+  getAllScheduledNotificationsAsync: () => Promise<NotificationRequest[]>;
   scheduleNotificationAsync: (
     identifier: string,
     notificationContent: NotificationContentInput,