+++ tools/expotools/src/Changelogs.ts
@@ -1,4 +1,5 @@
 import fs from 'fs-extra';
+import semver from 'semver';
 
 import * as Markdown from './Markdown';
 
@@ -29,6 +30,14 @@ export type ChangelogEntry = {
 };
 
 /**
+ * Type of the objects representing changelog entries.
+ */
+export type ChangelogChanges = {
+  totalCount: number;
+  versions: Record<string, Partial<Record<ChangeType, string[]>>>;
+};
+
+/**
  * Enum with changelog sections that are commonly used by us.
  */
 export enum ChangeType {
@@ -41,6 +50,9 @@ export enum ChangeType {
  * Heading name for unpublished changes.
  */
 export const UNPUBLISHED_VERSION_NAME = 'master';
+
+export const VERSION_EMPTY_PARAGRAPH_TEXT =
+  '*This version does not introduce any user-facing changes.*';
 
 /**
  * Depth of headings that mean the version containing following changes.
@@ -63,14 +75,106 @@ export class Changelog {
     this.filePath = filePath;
   }
 
+  /**
+   * Resolves to `true` if changelog file exists, `false` otherwise.
+   */
+  async fileExistsAsync(): Promise<boolean> {
+    return await fs.pathExists(this.filePath);
+  }
+
+  /**
+   * Lexifies changelog content and returns resulting tokens.
+   */
   async getTokensAsync(): Promise<Markdown.Tokens> {
     if (!this.tokens) {
-      await fs.access(this.filePath, fs.constants.R_OK);
-      this.tokens = Markdown.lexify(await fs.readFile(this.filePath, 'utf8'));
+      try {
+        this.tokens = Markdown.lexify(await fs.readFile(this.filePath, 'utf8'));
+      } catch (error) {
+        this.tokens = [];
+      }
     }
     return this.tokens;
   }
 
+  /**
+   * Reads versions headers, collects those versions and returns them.
+   */
+  async getVersionsAsync(): Promise<string[]> {
+    const tokens = await this.getTokensAsync();
+    const versionTokens = tokens.filter(
+      (token) => token.type === Markdown.TokenType.HEADING && token.depth === VERSION_HEADING_DEPTH
+    ) as Markdown.HeadingToken[];
+
+    return versionTokens.map((token) => token.text);
+  }
+
+  /**
+   * Returns the last version in changelog.
+   */
+  async getLastPublishedVersionAsync(): Promise<string | null> {
+    const versions = await this.getVersionsAsync();
+    return versions.find((version) => semver.valid(version)) ?? null;
+  }
+
+  /**
+   * Reads changes between two given versions and returns them in JS object format.
+   * If called without params, then only unpublished changes are returned.
+   */
+  async getChangesAsync(
+    fromVersion?: string,
+    toVersion: string = UNPUBLISHED_VERSION_NAME
+  ): Promise<ChangelogChanges> {
+    const tokens = await this.getTokensAsync();
+    const versions: ChangelogChanges['versions'] = {};
+    const changes: ChangelogChanges = { totalCount: 0, versions };
+
+    let currentVersion: string | null = null;
+    let currentSection: string | null = null;
+
+    for (let i = 0; i < tokens.length; i++) {
+      const token = tokens[i];
+
+      if (token.type === Markdown.TokenType.HEADING) {
+        if (token.depth === VERSION_HEADING_DEPTH) {
+          if (token.text !== toVersion && (!fromVersion || token.text === fromVersion)) {
+            // We've iterated over everything we needed, stop the loop.
+            break;
+          }
+
+          currentVersion = token.text === UNPUBLISHED_VERSION_NAME ? 'unpublished' : token.text;
+          currentSection = null;
+
+          if (!versions[currentVersion]) {
+            versions[currentVersion] = {};
+          }
+        } else if (currentVersion && token.depth === CHANGE_TYPE_HEADING_DEPTH) {
+          currentSection = token.text;
+
+          if (!versions[currentVersion][currentSection]) {
+            versions[currentVersion][currentSection] = [];
+          }
+        }
+        continue;
+      }
+
+      if (currentVersion && currentSection && token.type === Markdown.TokenType.LIST_ITEM_START) {
+        i++;
+        for (; tokens[i].type !== Markdown.TokenType.LIST_ITEM_END; i++) {
+          const token = tokens[i] as Markdown.TextToken;
+
+          if (token.text) {
+            changes.totalCount++;
+            versions[currentVersion][currentSection].push(token.text);
+          }
+        }
+      }
+    }
+    return changes;
+  }
+
+  /**
+   * Inserts given entry to changelog.
+   */
   async addChangeAsync(entry: ChangelogEntry): Promise<void> {
     const tokens = [...(await this.getTokensAsync())];
     const sectionIndex = tokens.findIndex(
@@ -150,6 +254,74 @@ export class Changelog {
 
     throw new Error(`Cound't find '${entry.type}' section.`);
   }
+
+  /**
+   * Renames header of unpublished changes to given version and adds new section with unpublished changes on top.
+   */
+  async cutOffAsync(version: string): Promise<void> {
+    const tokens = [...(await this.getTokensAsync())];
+    const firstVersionHeadingIndex = tokens.findIndex(isVersionToken);
+    const newSectionTokens: Markdown.Tokens = [
+      {
+        type: Markdown.TokenType.HEADING,
+        depth: VERSION_HEADING_DEPTH,
+        text: UNPUBLISHED_VERSION_NAME,
+      },
+      {
+        type: Markdown.TokenType.HEADING,
+        depth: CHANGE_TYPE_HEADING_DEPTH,
+        text: ChangeType.BREAKING_CHANGES,
+      },
+      {
+        type: Markdown.TokenType.HEADING,
+        depth: CHANGE_TYPE_HEADING_DEPTH,
+        text: ChangeType.NEW_FEATURES,
+      },
+      {
+        type: Markdown.TokenType.HEADING,
+        depth: CHANGE_TYPE_HEADING_DEPTH,
+        text: ChangeType.BUG_FIXES,
+      },
+    ];
+
+    if (firstVersionHeadingIndex !== -1) {
+      // Set version of the first found version header.
+      (tokens[firstVersionHeadingIndex] as Markdown.HeadingToken).text = version;
+
+      // Clean up empty sections.
+      let i = firstVersionHeadingIndex + 1;
+      while (i < tokens.length && !isVersionToken(tokens[i])) {
+        // Remove change type token if its section is empty - when it is followed by another heading token.
+        if (isChangeTypeToken(tokens[i])) {
+          const nextToken = tokens[i + 1];
+          if (!nextToken || isChangeTypeToken(nextToken) || isVersionToken(nextToken)) {
+            tokens.splice(i, 1);
+            continue;
+          }
+        }
+        i++;
+      }
+
+      // `i` stayed the same after removing empty change type sections, so the entire version is empty.
+      // Let's put an information that this version doesn't contain any user-facing changes.
+      if (i === firstVersionHeadingIndex + 1) {
+        tokens.splice(i, 0, {
+          type: Markdown.TokenType.PARAGRAPH,
+          text: VERSION_EMPTY_PARAGRAPH_TEXT,
+        });
+      }
+    }
+
+    // Insert new tokens before first version header.
+    tokens.splice(firstVersionHeadingIndex, 0, ...newSectionTokens);
+
+    // Parse tokens and write result to the file.
+    await fs.outputFile(this.filePath, Markdown.parse(tokens));
+
+    // Reset cached tokens as we just modified the file.
+    // We could use an array with new tokens here, but just for safety, let them be reloaded.
+    this.tokens = null;
+  }
 }
 
 /**
@@ -157,4 +329,22 @@ export class Changelog {
  */
 export function loadFrom(path: string): Changelog {
   return new Changelog(path);
+}
+
+/**
+ * Checks whether given token is interpreted as a token with a version.
+ */
+function isVersionToken(token: Markdown.Token): boolean {
+  return (
+    token && token.type === Markdown.TokenType.HEADING && token.depth === VERSION_HEADING_DEPTH
+  );
+}
+
+/**
+ * Checks whether given token is interpreted as a token with a change type.
+ */
+function isChangeTypeToken(token: Markdown.Token): boolean {
+  return (
+    token && token.type === Markdown.TokenType.HEADING && token.depth === CHANGE_TYPE_HEADING_DEPTH
+  );
 }